# [compile] optimize reduce nesting