--- a/compile.go
+++ b/compile.go
@@ -12,6 +12,7 @@ const (
 	Optimize        Option = "optimize" // switch all optimizations
 	Reordering      Option = "reordering"
 	FastEvaluation  Option = "fast_evaluation"
+	ReduceNesting   Option = "reduce_nesting"
 	ConstantFolding Option = "constant_folding"
 
 	Debug                 Option = "debug"
@@ -19,7 +20,17 @@ const (
 	AllowUnknownSelectors Option = "allow_unknown_selectors"
 )
 
-var AllOptimizations = []Option{Reordering, FastEvaluation, ConstantFolding}
+type optimizer struct {
+	name     Option
+	optimize func(config *CompileConfig, root *astNode)
+}
+
+var optimizers = []optimizer{
+	{name: ConstantFolding, optimize: optimizeConstantFolding},
+	{name: ReduceNesting, optimize: optimizeReduceNesting},
+	{name: FastEvaluation, optimize: optimizeFastEvaluation},
+	{name: Reordering, optimize: optimizeReordering},
+}
 
 func CopyCompileConfig(origin *CompileConfig) *CompileConfig {
 	conf := NewCompileConfig()
@@ -55,10 +66,15 @@ var (
 	}
 	Optimizations = func(enable bool, opts ...Option) CompileOption {
 		return func(c *CompileConfig) {
-			if len(opts) == 0 || opts[0] == Optimize {
-				opts = AllOptimizations
+			if len(opts) == 0 || (len(opts) == 1 && opts[0] == Optimize) {
+				for _, o := range optimizers {
+					c.CompileOptions[o.name] = enable
+				}
+				return
 			}
+
 			for _, opt := range opts {
+				// todo check opt
 				c.CompileOptions[opt] = enable
 			}
 		}
@@ -161,8 +177,10 @@ func Compile(originConf *CompileConfig, exprStr string) (*Expr, error) {
 
 func optimize(cc *CompileConfig, root *astNode) {
 	if enabled, exist := cc.CompileOptions[ConstantFolding]; enabled || !exist {
-		_ = optimizeConstantFolding(cc, root)
+		optimizeConstantFolding(cc, root)
 	}
+
+	optimizeReduceNesting(cc, root)
 
 	if enabled, exist := cc.CompileOptions[FastEvaluation]; enabled || !exist {
 		optimizeFastEvaluation(cc, root)
@@ -171,6 +189,38 @@ func optimize(cc *CompileConfig, root *astNode) {
 	if enabled, exist := cc.CompileOptions[Reordering]; enabled || !exist {
 		optimizeReordering(cc, root)
 	}
+}
+
+func optimizeReduceNesting(cc *CompileConfig, root *astNode) {
+	for _, child := range root.children {
+		optimizeReduceNesting(cc, child)
+	}
+
+	n := root.node
+	// todo expand to operators with associative property
+	if !isBoolOpNode(n) {
+		return
+	}
+
+	var children []*astNode
+	rootOpType := isAndOpNode(n)
+	for _, child := range root.children {
+		cn := child.node
+		if typ := cn.getNodeType(); typ == constant || typ == selector {
+			children = append(children, child)
+			continue
+		}
+		if !isBoolOpNode(cn) {
+			return
+		}
+		if isAndOpNode(cn) == rootOpType {
+			children = append(children, child.children...)
+			continue
+		}
+		return
+	}
+
+	root.children = children
 }
 
 func isBoolOpNode(n *node) bool {
@@ -282,18 +332,15 @@ func calculateNodeCosts(conf *CompileConfig, root *astNode) {
 	root.cost = int(cost)
 }
 
-func optimizeConstantFolding(cc *CompileConfig, root *astNode) error {
+func optimizeConstantFolding(cc *CompileConfig, root *astNode) {
 	for _, child := range root.children {
-		err := optimizeConstantFolding(cc, child)
-		if err != nil {
-			return err
-		}
+		optimizeConstantFolding(cc, child)
 	}
 
 	n := root.node
 	stateless, fn := isStatelessOp(cc, n)
 	if !stateless {
-		return nil
+		return
 	}
 
 	if isBoolOpNode(n) {
@@ -304,7 +351,7 @@ func optimizeConstantFolding(cc *CompileConfig, root *astNode) error {
 
 			b, ok := child.node.value.(bool)
 			if !ok {
-				return ParamTypeError(n.value.(string), "bool", child.node.value)
+				return
 			}
 
 			if (b && isOrOpNode(n)) || (!b && isAndOpNode(n)) {
@@ -313,7 +360,7 @@ func optimizeConstantFolding(cc *CompileConfig, root *astNode) error {
 					value: b,
 				}
 				root.children = nil
-				return nil
+				return
 			}
 		}
 	}
@@ -321,21 +368,21 @@ func optimizeConstantFolding(cc *CompileConfig, root *astNode) error {
 	params := make([]Value, len(root.children))
 	for i, child := range root.children {
 		if child.node.getNodeType() != constant {
-			return nil
+			return
 		}
 		params[i] = child.node.value
 	}
 
 	res, err := fn(nil, params)
 	if err != nil {
-		return err
+		return
 	}
 	root.children = nil
 	root.node = &node{
 		flag:  constant,
 		value: res,
 	}
-	return nil
+	return
 }
 
 func isStatelessOp(c *CompileConfig, n *node) (bool, Operator) {